pyproject.toml
[project]
name = "a2asimpeltask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[grpc]>=0.2.14",
    "ruff>=0.12.4",
    "uvicorn>=0.35.0",
]
